Split similarity into features and index.
